# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the VyoLayer API documentation.",
    "title": "VyoLayer API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "API Support",
      "url": "http://www.swagger.io/support",
      "email": "support@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0"
  },
  "host": "localhost:6999",
  "basePath": "/api/v1",
  "paths": {
...
    "/organizations/invitations/accept": {
      "post": {
        "description": "Accept an invitation to join an organization",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "Accept an invitation",
        "parameters": [
          {
            "type": "string",
            "description": "Invitation token",
            "name": "org-invite-token",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/invitations/pending": {
      "get": {
        "description": "Get all pending invitations for the authenticated user's email",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "Get pending invitations for current user",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.OrganizationMemberInvitationDTO"
                      }
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/invitations/{invitationId}": {
      "delete": {
        "description": "Cancel/delete an invitation",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "Cancel an invitation",
        "parameters": [
          {
            "type": "string",
            "description": "Invitation ID",
            "name": "invitationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/response.SuccessResponse"
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
    "/organizations/{orgId}/invitations": {
      "get": {
        "description": "Get all invitations for an organization",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "List organization invitations",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/definitions/dto.OrganizationMemberInvitationDTO"
                      }
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      },
      "post": {
        "description": "Create a new invitation to join an organization",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "Create an organization member invitation",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "description": "Invitation details",
            "name": "request",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dto.CreateInvitationRequestDTO"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationMemberInvitationDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "409": {
            "description": "Conflict",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "500": {
            "description": "Internal Server Error",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
    "/organizations/{orgId}/invitations/{invitationId}/resend": {
      "post": {
        "description": "Regenerate token and reset expiry for a pending invitation",
        "produces": ["application/json"],
        "tags": ["Organization Invitations"],
        "summary": "Resend an invitation",
        "parameters": [
          {
            "type": "string",
            "description": "Organization ID",
            "name": "orgId",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "Invitation ID",
            "name": "invitationId",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.SuccessResponse"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/dto.OrganizationMemberInvitationDTO"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "403": {
            "description": "Forbidden",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          },
          "404": {
            "description": "Not Found",
            "schema": {
              "$ref": "#/definitions/response.ErrorResponse"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "dto.CreateInvitationRequestDTO": {
      "type": "object",
      "required": ["email"],
      "properties": {
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "roleIds": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["[\"org_role_123\"]"]
        }
      }
    },
...
    "dto.OrganizationMemberInvitationDTO": {
      "type": "object",
      "properties": {
        "acceptedAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "email": {
          "type": "string",
          "example": "john@example.com"
        },
        "expiredAt": {
          "type": "string",
          "example": "2023-01-08T00:00:00Z"
        },
        "id": {
          "type": "string",
          "example": "org_invitation_550e8400-e29b-41d4-a716-446655440000"
        },
        "invitedAt": {
          "type": "string",
          "example": "2023-01-01T00:00:00Z"
        },
        "invitedBy": {
          "type": "string",
          "example": "org_member_550e8400-e29b-41d4-a716-446655440000"
        },
        "isAccepted": {
          "type": "boolean",
          "example": false
        },
        "isPending": {
          "type": "boolean",
          "example": true
        },
        "organizationId": {
          "type": "string",
          "example": "org_550e8400-e29b-41d4-a716-446655440000"
        },
        "roleIds": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "example": ["[\"org_role_123\"]"]
        }
      }
    },
...
    "response.ErrorDetail": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string",
          "example": "ERR_VALIDATION_FAILED"
        },
        "details": {},
        "message": {
          "type": "string",
          "example": "Validation failed"
        },
        "metadata": {
          "type": "object",
          "additionalProperties": true
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        }
      }
    },
    "response.ErrorResponse": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/response.ErrorDetail"
        },
        "statusCode": {
          "type": "integer",
          "example": 400
        },
        "success": {
          "type": "boolean",
          "example": false
        }
      }
    },
    "response.Meta": {
      "type": "object",
      "properties": {
        "limit": {
          "type": "integer",
          "example": 10
        },
        "page": {
          "type": "integer",
          "example": 1
        },
        "requestId": {
          "type": "string",
          "example": "req-123456"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-02-08T18:30:00Z"
        },
        "total": {
          "type": "integer",
          "example": 100
        },
        "totalPages": {
          "type": "integer",
          "example": 10
        }
      }
    },
    "response.SuccessResponse": {
      "type": "object",
      "properties": {
        "data": {},
        "message": {
          "type": "string",
          "example": "Operation successful"
        },
        "meta": {
          "$ref": "#/definitions/response.Meta"
        },
        "statusCode": {
          "type": "integer",
          "example": 200
        },
        "success": {
          "type": "boolean",
          "example": true
        }
      }
    }
  }
}